+++ src/components/ui/SearchInput/index.tsx
@@ -1,7 +1,7 @@
 import { memo, SyntheticEvent } from 'react';
 import searchIcon from '@assets/icons/search.svg';
 
-import { configPlaceholder } from '@/helpers/configireSearchPlaceholder';
+import { SearchPath, SearchPlaceholder } from '@/constants/textConstant';
 
 import { Container, Icon, SearchInputContainer } from './searchInput.styled';
 
@@ -10,15 +10,20 @@ interface SearchInputProps {
 	searchValue: string;
 	onChangeHandler: (e: SyntheticEvent) => void;
 }
+
 export const SearchInput = memo(
-	({ searchPath, searchValue, onChangeHandler }: SearchInputProps) => (
-		<Container>
-			<Icon src={searchIcon} alt="search" />
-			<SearchInputContainer
-				placeholder={configPlaceholder(searchPath)}
-				value={searchValue}
-				onChange={onChangeHandler}
-			/>
-		</Container>
-	)
+	({ searchPath, searchValue, onChangeHandler }: SearchInputProps) => {
+		const configPlaceholder = (value: string): string =>
+			value === SearchPath.users.toString() ? SearchPlaceholder.users : SearchPlaceholder.tweets;
+		return (
+			<Container>
+				<Icon src={searchIcon} alt="search" />
+				<SearchInputContainer
+					placeholder={configPlaceholder(searchPath)}
+					value={searchValue}
+					onChange={onChangeHandler}
+				/>
+			</Container>
+		);
+	}
 );
